pc-llm-service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
httpx==0.27.2
pydantic==2.9.2
paho-mqtt==2.1.0
opencv-python-headless==4.10.0.84
pyyaml==6.0.2
numpy==1.26.4
Pillow==10.4.0
sentence-transformers==3.1.1
chromadb==0.5.15
